Trait futures_codec::Decoder

pub trait Decoder {
    type Item;
    type Error: From<Error>;
    fn decode(
        &mut self, 
        src: &mut BytesMut
    ) -> Result<Option<Self::Item>, Self::Error>;

    fn decode_eof(
        &mut self, 
        src: &mut BytesMut
    ) -> Result<Option<Self::Item>, Self::Error> { ... }
}

Decoding of frames via buffers, for use with FramedRead.

Associated Types

type Item

The type of items returned by decode

type Error: From<Error>

The type of decoding errors.

Loading content...
Required methods

fn decode(
&mut self,
src: &mut BytesMut
) -> Result<Option<Self::Item>, Self::Error>

Decode an item from the src BytesMut into an item

Loading content...
Provided methods

fn decode_eof(
&mut self,
src: &mut BytesMut
) -> Result<Option<Self::Item>, Self::Error>

Called when the input stream reaches EOF, signaling a last attempt to decode

Notes

The default implementation of this method invokes the Decoder::decode method.

Loading content...